:recycle: refactor: remove deprecate
* ♻️  refactor: remove deprecate class AgentRuntime
* ♻️  refactor: fix circular dependence from AgentRuntime
* :white_check_mark: test: pass tests
