+++ src/libs/agent-runtime/AgentRuntime.ts
@@ -1,27 +1,23 @@
 import { ClientOptions } from 'openai';
 
 import type { TracePayload } from '@/const/trace';
-import {
-  ChatCompetitionOptions,
-  ChatStreamPayload,
-  LobeAnthropicAI,
-  LobeAzureOpenAI,
-  LobeBedrockAI,
-  LobeGoogleAI,
-  LobeGroq,
-  LobeMistralAI,
-  LobeMoonshotAI,
-  LobeOllamaAI,
-  LobeOpenAI,
-  LobeOpenRouterAI,
-  LobePerplexityAI,
-  LobeRuntimeAI,
-  LobeTogetherAI,
-  LobeZeroOneAI,
-  LobeZhipuAI,
-  ModelProvider,
-} from '@/libs/agent-runtime';
-import { LobeBedrockAIParams } from '@/libs/agent-runtime/bedrock';
+import { LobeAnthropicAI } from '@/libs/agent-runtime/anthropic';
+import { LobeAzureOpenAI } from '@/libs/agent-runtime/azureOpenai';
+import { LobeBedrockAI, LobeBedrockAIParams } from '@/libs/agent-runtime/bedrock';
+import { LobeGoogleAI } from '@/libs/agent-runtime/google';
+import { LobeGroq } from '@/libs/agent-runtime/groq';
+import { LobeMistralAI } from '@/libs/agent-runtime/mistral';
+import { LobeMoonshotAI } from '@/libs/agent-runtime/moonshot';
+import { LobeOllamaAI } from '@/libs/agent-runtime/ollama';
+import { LobeOpenAI } from '@/libs/agent-runtime/openai';
+import { LobeOpenRouterAI } from '@/libs/agent-runtime/openrouter';
+import { LobePerplexityAI } from '@/libs/agent-runtime/perplexity';
+import { LobeTogetherAI } from '@/libs/agent-runtime/togetherai';
+import { LobeZeroOneAI } from '@/libs/agent-runtime/zeroone';
+import { LobeZhipuAI } from '@/libs/agent-runtime/zhipu';
+
+import { LobeRuntimeAI } from './BaseAI';
+import { ChatCompetitionOptions, ChatStreamPayload, ModelProvider } from './types';
 
 export interface AgentChatOptions {
   enableTrace?: boolean;
