+++ src/libs/agent-runtime/index.ts
@@ -1,3 +1,4 @@
+export { default as AgentRuntime } from './AgentRuntime';
 export { LobeAnthropicAI } from './anthropic';
 export { LobeAzureOpenAI } from './azureOpenai';
 export * from './BaseAI';
