+++ src/services/__tests__/chat.test.ts
@@ -21,7 +21,7 @@ import {
   LobeZhipuAI,
   ModelProvider,
 } from '@/libs/agent-runtime';
-import AgentRuntimeLib from '@/libs/agent-runtime/AgentRuntime';
+import { AgentRuntime } from '@/libs/agent-runtime';
 import { useFileStore } from '@/store/file';
 import { GlobalStore } from '@/store/global';
 import {
@@ -697,7 +697,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.OpenAI, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeOpenAI);
         expect(runtime['_runtime'].baseURL).toBe('user-openai-endpoint');
       });
@@ -715,7 +715,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Azure, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeAzureOpenAI);
       });
 
@@ -730,7 +730,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Google, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeGoogleAI);
       });
 
@@ -745,7 +745,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Moonshot, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeMoonshotAI);
       });
 
@@ -762,7 +762,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Bedrock, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeBedrockAI);
       });
 
@@ -777,7 +777,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Ollama, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeOllamaAI);
       });
 
@@ -792,7 +792,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Perplexity, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobePerplexityAI);
       });
 
@@ -807,7 +807,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Anthropic, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeAnthropicAI);
       });
 
@@ -822,7 +822,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Mistral, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeMistralAI);
       });
 
@@ -837,7 +837,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.OpenRouter, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeOpenRouterAI);
       });
 
@@ -852,7 +852,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.TogetherAI, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeTogetherAI);
       });
 
@@ -867,7 +867,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.ZeroOne, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeZeroOneAI);
       });
 
@@ -882,7 +882,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.Groq, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeGroq);
       });
 
@@ -902,7 +902,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as any as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore('unknown' as ModelProvider, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeOpenAI);
       });
 
@@ -929,7 +929,7 @@ describe('AgentRuntimeOnClient', () => {
           },
         } as GlobalSettingsState) as unknown as GlobalStore;
         const runtime = await initializeWithClientStore(ModelProvider.ZhiPu, {});
-        expect(runtime).toBeInstanceOf(AgentRuntimeLib);
+        expect(runtime).toBeInstanceOf(AgentRuntime);
         expect(runtime['_runtime']).toBeInstanceOf(LobeZhipuAI);
       });
     });
